feat(1900): Adjustments after code review
